Fix: Estabelecimento field CNPJ data type fix to long.

+++ Data/DTO/EstabelecimentoDTO/CreateEstabelecimentoDTO.cs
@@ -6,7 +6,7 @@ namespace API_Fase2.Data.DTO.EstabelecimentoDTO
     {
 
         [Required]
-        public int Cnpj { get; set; }
+        public long Cnpj { get; set; }
 
         [Required(ErrorMessage = "O campo NomeEstabelecimento é obrigatório.")]
         [StringLength(50, ErrorMessage = "O tamanho máximo do campo é 50 caracteres.")]
